Add variable transformation
