+++ Makefile
@@ -85,7 +85,7 @@ wisb_flsim3: $(IO_SRCS) $(WISB_SRCS) $(FSIM3_SRCS)
 #
 
 ws2st: $(wildcard src/ws2st/*.cpp)
-	$(CXX) $(wildcard src/ws2st/*.cpp) $(CLFAGS) -g -lclang-cpp -lLLVM-16 -o $(BUILD_PATH)/refactor
+	$(CXX) $(wildcard src/ws2st/*.cpp) $(wildcard src/ws2st/variables/*.cpp) $(CLFAGS) -g -lclang-cpp -lLLVM-16 -o $(BUILD_PATH)/refactor
 
 dsl_flsim2: ws2st $(FSIM_SRCS)
 	$(BUILD_PATH)/refactor ./src/examples/ca_fluid_sim -- -I/usr/lib64/clang/16/include -DDIM=2 -fopenmp
