+++ src/ws2st/reg_forces.hpp
@@ -10,6 +10,7 @@
 #pragma once
 
 #include "common.hpp"
+#include "variables/variables.hpp"
 
 namespace wash {
